src/bigint bi_read_dec: negative not working out; let's rewrite read_udec
The whole "convert positive to negative thing" isn't working out well,
and is almost-but-not-quite duplicating a lot of code. Perhaps it's
better to change _udec instead to have a parameter to tell it to
subtract, instead of add, and see if that gets us back to something
closer to a single case for positive and negative in the _dec routine.