src/bigint bi_x10: Do signed multiply as well (it always did)
Rename bi_x10u to bi_x10 and add tests confirming signed
multiplication and demonstrating some minimum and maximum values we
can use for it. Better characterize it in the comments, including
properly explaining a heuristic for min/max values, though we'll
continue using a looser one when calling it from the decimal digit
reader.

Also properly document and start using the bufptr₁ documentation
convention for buffers indexed from 1 (where the pointer points to the
byte just before the buffer).

Also rename bi_read_udec to bi_read_decdigits and start tweaking it in
preparation for having it do a signed read.