src/bigint bi_read_dec: Read an ASCII decimal number
I'm not utterly convinced of this code's reliability: during development
serious bugs were missed for some time due to bad choices of test cases.
Probably someone should go over the code and test cases carefully and
figure out if there is coverage we need but are missing.

This is three routines:
- bi_x10 (55 bytes): in-place multiply by 10
- bi_read_decdigits (78 bytes): digit reading (sign specified separately)
- bi_read_dec (118 bytes): main routine; handles sign characters and
  normalization, and '0' as a special case.

These routines don't validate their input (e.g., for invalid characters)
because they're intended to be used by a more general parser that would be
calling bi_read_dec only if it had already decided that the string was a
decimal number, and not something else.

This is large (251 bytes total, or about 7% of available RAM on 4K system,
ouch!) but it's hard to see what can be done about that; base 10→16
conversion is just complex. It might help later to see how much of this
could be shared with a floating point read routine.

It's also not quick, though there are a few optimizations (mentioned in the
comments) that could be done, some of which are not too hard but expand the
code and others which involve things like adding self-modifying code.

There's certainly some low-hanging refactoring available to be done here;
one obvious example is to extract the repeated 16-bit pointer increment
code to a macro.

Better optimizing the normalization to avoid allocating a temporary
buffer that's usually to long and then copying it out to the final
buffer turns out to be very difficult; I don't know if there's an
easier or faster way than doing the actual conversion itself. This
wouldn't be much of an optimization overall, anyway, becuase most of
the time is taken by the repeated multiplies by 10.

Currently all memory management (sizing and allocation of temporary
buffers) needs to be done by the bi_read_dec caller. Once we have some sort
of memory management available, the temporary buffers at least should
probably be allocated by these routines, since for efficiency bi_read_dec
already calculates the sizes of them anyway. It would be good to make the
API for that work in a way that a different allocation system could be
fairly easily substituted in case one wants to use the bigint library in
other, different, programs.

The use of variables bufNptr and the like really wants to be improved, but
it's not totally obvious how to do that right now, and having a memory
allocation system will affect this significantly anyway, so that should
wait until after we've got an alloc system in place. When writing the alloc
system, how this would use it conveniently should be kept in mind.

The unit test code also really wants a bunch of cleanup and refactoring of
the buffer allocation and bounds checking, but again, that is best held off
until we see what the memory allocation system looks like. The alloc system
might even offer its own bounds checks that the unit tests could use.

Personal note: This was a _lot_ of work: the 300+ lines of code and
comments were developed over 30 commits and five days with many false
turns and restarts. It's my first major piece of non-trivial 6502 code
and was very much a learning experience.