src/mc68/pmon regscmd: Start on error message
Message is currently `?nn c` where `nn` is supposed to be the offset from
the beginning of the command line buffer where the offending character was
found (but this is not yet implemented) and `c` is the offending character
(which does work).

Also had to move the show code from afer the args code to before to bring
some branches back into range, though for the show we now had to change
some BSRs to JSRs. (That probably would have happened anyway when we moved
prchval to a new home, if this is not the only place we use it. Or maybe it
will just get moved inline when we build a table to give the register
names, sizes and locations. That could be used for the parsing, too!)

Also renamed prcheqval to prchval, since we changed it a while back to
print 's1234' rather than 's=1234' to crunch this down to fit into 32 char
lines. (Obviously this will still wrap later on 8080, though.)